FIX :: features 폴더 구조화 변경
