--- src/features/auth/login/login.props.ts
@@ -1,5 +1,0 @@
-import { LoginType } from 'types/auth/auth.type';
-
-export interface LoginFeatureType {
-  loginData: LoginType;
-}
