--- src/features/auth/school/school.props.ts
@@ -1,3 +1,0 @@
-export interface SchoolFeatureType {
-  searchWord: string;
-}
