+++ src/features/home/detail.feature.ts
