--- src/features/home/petition/answer/answer.props.ts
@@ -1,7 +1,0 @@
-import { Dispatch, SetStateAction } from 'react';
-
-export interface AnswerFeatureType {
-  detailId: number;
-  setComment: Dispatch<SetStateAction<string>>;
-  comment: string;
-}
