--- src/features/home/petition/comment/comment.props.ts
@@ -1,7 +1,0 @@
-import { Dispatch, SetStateAction } from 'react';
-
-export interface CommentFeatureType {
-  detailId: number;
-  setComment: Dispatch<SetStateAction<string>>;
-  comment: string;
-}
