+++ src/types/common/button.type.ts
@@ -1,4 +1,4 @@
-import React, { ButtonHTMLAttributes } from 'react';
+import { ButtonHTMLAttributes } from 'react';
 
 export interface ButtonPropsType
   extends ButtonHTMLAttributes<HTMLButtonElement> {
